Civ_5_Mod_Maker
This is another Java project that I worked on. It converts input from
the GUI to data on an XML file that can be used in Civilization V. It is
not the cleanest code I have written and there are still some bugs, but
it gets the job done most of the time.